requirements.txt
ultralytics
streamlit
pillow
opencv-python-headless
easyocr
